feat(sentry): Do not capture http requests with response code… (#15734)
A lot of these are expected and not an indicator of actual problems. This especially creates a lot of noise for on-premise installs. These requests are also not very helpful when grouped so lets just not record them at all.

(cherry picked from commit 3a343fdaba7a51bc8aed5b1f244f007a6e557317)