ref(ui): Check first event sent via API in Issues Overview (#16096)
Previously, the issues page would use the list of all projects (organization.projects) to check if any selected projects had sent a first event. This meant that the issues page was blocked on the loading of all projects. Now it is changed to make use of the sent-first-event/ API call which can take in is_member or project ids as query parameters. Also fixed up some tests that were not correctly waiting for the right async calls.